+++ src/renderer-moderngl/DisplayObject.hpp
@@ -94,7 +94,7 @@ public:
 	virtual void killMe();
 	void setTween(TweenSlot slot, easing_ptr easing, float from, float to, float time, int on_end_ref, int on_change_ref);
 	void cancelTween(TweenSlot slot);
-	virtual void onKeyframe(int nb_keyframes);
+	virtual void onKeyframe(float nb_keyframes);
 };
 
 extern int donb;
