+++ src/renderer-moderngl/Interfaces.cpp
@@ -67,7 +67,7 @@ void IRealtime::killMe() {
 	}
 }
 
-void interface_realtime(int nb_keyframes) {
+void interface_realtime(float nb_keyframes) {
 	realtime_executing = true;
 	for (auto it = realtime_list.begin(); it != realtime_list.end();) {
 		(*it)->onKeyframe(nb_keyframes);
